refactor(tests): reorganize and modularize test structure for improved maintainability and clarity, including model serialization, configuration, authentication, and middleware tests
