+++ crates/dbx-api/src/tests/mod.rs
@@ -1,0 +1,10 @@
+//! DBX API test modules
+//!
+//! Tests are organized by functionality for maintainability and clarity.
+//! Each module focuses on a specific aspect of the API implementation.
+
+/// Shared test utilities and helper functions
+pub mod utils;
+
+/// Model serialization and validation tests  
+mod model_tests;
